+++ src/Mailcode/Translator/Syntax/ApacheVelocity.php
@@ -12,6 +12,7 @@ declare(strict_types=1);
 namespace Mailcode;
 
 use Mailcode\Interfaces\Commands\EncodableInterface;
+use Mailcode\Interfaces\Commands\Validation\DecryptInterface;
 
 /**
  * Abstract base class for apache velocity command translation classes.
@@ -49,16 +50,16 @@ abstract class Mailcode_Translator_Syntax_ApacheVelocity extends Mailcode_Transl
     }
 
     /**
-    * Filters the string for use in an Apache Velocity (Java)
-    * regex string: escapes all special characters.
-    *
-    * Velocity does its own escaping, so no need to escape special
-    * characters as if they were a javascript string.
-    *
-    * @param string $string
-    * @return string
-    */
-    public function filterRegexString(string $string) : string
+     * Filters the string for use in an Apache Velocity (Java)
+     * regex string: escapes all special characters.
+     *
+     * Velocity does its own escaping, so no need to escape special
+     * characters as if they were a javascript string.
+     *
+     * @param string $string
+     * @return string
+     */
+    public function filterRegexString(string $string): string
     {
         // Special case: previously escaped quotes.
         // To avoid modifying them, we strip them out.
@@ -71,9 +72,8 @@ abstract class Mailcode_Translator_Syntax_ApacheVelocity extends Mailcode_Transl
 
         // All other special characters have to be escaped
         // with two backslashes.
-        foreach($this->regexSpecialChars as $char)
-        {
-            $string = str_replace($char, '\\'.$char, $string);
+        foreach ($this->regexSpecialChars as $char) {
+            $string = str_replace($char, '\\' . $char, $string);
         }
 
         // Restore the escaped quotes, which stay escaped
@@ -81,6 +81,11 @@ abstract class Mailcode_Translator_Syntax_ApacheVelocity extends Mailcode_Transl
         $string = str_replace('ESCQUOTE', '\\"', $string);
 
         return $string;
+    }
+
+    protected function hasVariableEncodings(Mailcode_Commands_Command $command) : bool
+    {
+        return $command instanceof EncodableInterface && $command->hasActiveEncodings();
     }
 
     protected function renderVariableEncodings(Mailcode_Commands_Command $command, string $varName): string
@@ -99,8 +104,7 @@ abstract class Mailcode_Translator_Syntax_ApacheVelocity extends Mailcode_Transl
     {
         $varName = dollarize($varName);
 
-        if($absolute)
-        {
+        if ($absolute) {
             $varName = sprintf('${numeric.abs(%s)}', $varName);
         }
 
@@ -150,13 +154,13 @@ abstract class Mailcode_Translator_Syntax_ApacheVelocity extends Mailcode_Transl
         $result = $statement;
 
         foreach ($encodings as $encoding) {
-            $result = $this->renderEncoding($encoding, $result);
+            $result = $this->renderEncoding($encoding, $result, $command);
         }
 
         return $result;
     }
 
-    protected function renderEncoding(string $keyword, string $result): string
+    protected function renderEncoding(string $keyword, string $result, EncodableInterface $command): string
     {
         $template = $this->encodingTemplates[$keyword] ?? '%s';