+++ tests/testsuites/Commands/Types/ShowVarTests.php
@@ -4,6 +4,7 @@ declare(strict_types=1);
 
 namespace testsuites\Commands\Types;
 
+use Mailcode\Decrypt\DecryptSettings;
 use Mailcode\Mailcode;
 use Mailcode\Mailcode_Commands_Command;
 use Mailcode\Mailcode_Commands_Command_ShowVariable;
@@ -41,20 +42,38 @@ final class ShowVarTests extends MailcodeTestCase
                 'code' => 0
             ),
             array(
-                'label' => 'With valid variable and another parameter',
+                'label' => 'With valid variable and another parameter, which is ignored',
                 'string' => '{showvar: $foo_bar "Text"}',
-                'valid' => false,
-                'code' => Mailcode_Commands_Command_ShowVariable::VALIDATION_TOO_MANY_PARAMETERS
+                'valid' => true,
+                'code' => 0
+            ),
+            array(
+                'label' => 'With valid variable and default decryption key',
+                'string' => '{showvar: $foo_bar decrypt=""}',
+                'valid' => true,
+                'code' => 0
+            ),
+            array(
+                'label' => 'With valid variable and custom decryption key',
+                'string' => '{showvar: $foo_bar "barfoo" decrypt="fookey"}',
+                'valid' => true,
+                'code' => 0
+            ),
+            array(
+                'label' => 'With valid variable and custom decryption key',
+                'string' => '{showvar: $foo_bar decrypt="fookey" idnencode:}',
+                'valid' => true,
+                'code' => 0
             ),
         );
-        
+
         $this->runCollectionTests($tests);
     }
-    
+
     public function test_getVariable() : void
     {
         $cmd = Mailcode_Factory::show()->var('foobar');
-        
+
         $this->assertEquals('$foobar', $cmd->getVariable()->getFullName());
         $this->assertEquals('$foobar', $cmd->getVariableName());
     }
@@ -65,16 +84,14 @@ final class ShowVarTests extends MailcodeTestCase
      */
     public function test_urlEncodeFromParams() : void
     {
-        $cmd = Mailcode::create()->parseString('{showvar: $FOO urlencode:}')->getFirstCommand();
+        $cmd = $this->getCommandFromString('{showvar: $FOO urlencode:}');
 
-        $this->assertInstanceOf(Mailcode_Commands_Command_ShowVariable::class, $cmd);
         $this->assertTrue($cmd->isURLEncoded());
     }
 
     public function test_urlEncodeFromSet() : void
     {
-        $cmd = Mailcode::create()->parseString('{showvar: $FOO}')->getFirstCommand();
-        $this->assertInstanceOf(Mailcode_Commands_Command_ShowVariable::class, $cmd);
+        $cmd = $this->getCommandFromString('{showvar: $FOO}');
 
         $cmd->setURLEncoding(true);
         $this->assertEquals('{showvar: $FOO urlencode:}', $cmd->getNormalized());
@@ -85,29 +102,95 @@ final class ShowVarTests extends MailcodeTestCase
 
     public function test_bothEncodeAndDecodeError() : void
     {
-        $cmd = Mailcode::create()->parseString('{showvar: $FOO urlencode: urldecode:}');
+        $mailcode = Mailcode::create()->parseString('{showvar: $FOO urlencode: urldecode:}');
 
-        $this->assertFalse($cmd->isValid());
-        $this->assertEquals(Mailcode_Commands_CommonConstants::VALIDATION_URL_DE_AND_ENCODE_ENABLED, $cmd->getFirstError()->getCode());
+        $this->assertFalse($mailcode->isValid());
+        $this->assertEquals(Mailcode_Commands_CommonConstants::VALIDATION_URL_DE_AND_ENCODE_ENABLED, $mailcode->getFirstError()->getCode());
     }
 
     public function test_idnEncodeFromParams() : void
     {
-        $cmd = Mailcode::create()->parseString('{showvar: $FOO idnencode:}')->getFirstCommand();
+        $cmd = $this->getCommandFromString('{showvar: $FOO idnencode:}');
 
-        $this->assertInstanceOf(Mailcode_Commands_Command_ShowVariable::class, $cmd);
         $this->assertTrue($cmd->isIDNEncoded());
     }
 
     public function test_idnEncodeFromSet() : void
     {
-        $cmd = Mailcode::create()->parseString('{showvar: $FOO}')->getFirstCommand();
-        $this->assertInstanceOf(Mailcode_Commands_Command_ShowVariable::class, $cmd);
+        $cmd = $this->getCommandFromString('{showvar: $FOO}');
 
         $cmd->setIDNEncoding(true);
         $this->assertEquals('{showvar: $FOO idnencode:}', $cmd->getNormalized());
 
         $cmd->setIDNEncoding(false);
         $this->assertEquals('{showvar: $FOO}', $cmd->getNormalized());
+    }
+
+    public function test_decryptDefault() : void
+    {
+        DecryptSettings::setDefaultKeyName('my-key');
+
+        $cmd = $this->getCommandFromString('{showvar: $FOO decrypt=""}');
+
+        $this->assertTrue($cmd->isDecryptionEnabled());
+        $this->assertEquals('my-key', $cmd->getDecryptionKeyName());
+        $this->assertEquals('{showvar: $FOO decrypt=""}', $cmd->getNormalized());
+    }
+
+    public function test_decryptCustom() : void
+    {
+        $cmd = $this->getCommandFromString('{showvar: $FOO decrypt="custom-key"}');
+
+        $this->assertTrue($cmd->isDecryptionEnabled());
+        $this->assertEquals('custom-key', $cmd->getDecryptionKeyName());
+        $this->assertEquals('{showvar: $FOO decrypt="custom-key"}', $cmd->getNormalized());
+    }
+
+    public function test_decryptMethods() : void
+    {
+        $cmd = $this->getCommandFromString('{showvar: $FOO}');
+
+        $cmd->enableDecryption();
+
+        $this->assertEquals('{showvar: $FOO decrypt="default"}', $cmd->getNormalized());
+
+        $cmd->disableDecryption();
+
+        $this->assertEquals('{showvar: $FOO}', $cmd->getNormalized());
+    }
+
+    public function test_decryptNoDefault() : void
+    {
+        $this->assertNull(DecryptSettings::getDefaultKeyName(), 'Precondition for the test is no default key.');
+
+        $mailcode = $this->getCommandFromString('{showvar: $FOO decrypt=""}');
+
+        $this->assertEmpty($mailcode->getDecryptionKeyName());
+    }
+
+    public function test_decryptDefaultKeyWhenDefault() : void
+    {
+        DecryptSettings::setDefaultKeyName('default-key');
+
+        $cmd = $this->getCommandFromString('{showvar: $FOO decrypt=""}');
+
+        $this->assertEquals('default-key', $cmd->getDecryptionKeyName());
+    }
+
+    public function test_decryptDefaultKeyWhenEmpty() : void
+    {
+        DecryptSettings::setDefaultKeyName('default-key');
+
+        $cmd = $this->getCommandFromString('{showvar: $FOO decrypt=""}');
+
+        $this->assertEquals('default-key', $cmd->getDecryptionKeyName());
+    }
+
+    private function getCommandFromString(string $command) : Mailcode_Commands_Command_ShowVariable
+    {
+        $cmd = Mailcode::create()->parseString($command)->getFirstCommand();
+        $this->assertInstanceOf(Mailcode_Commands_Command_ShowVariable::class, $cmd);
+
+        return $cmd;
     }
 }