+++ tests/testsuites/Translator/ApacheVelocity.php
@@ -12,7 +12,7 @@ final class Translator_ApacheVelocityTests extends VelocityTestCase
 
         $expected = '#if($FOO.BAR == 20 && $BARFOO == "Other value")';
 
-        $syntax = $this->translator->createSyntax('ApacheVelocity');
+        $syntax = $this->translator->createApacheVelocity();
 
         $result = $syntax->translateCommand($cmd);
 
@@ -25,7 +25,36 @@ final class Translator_ApacheVelocityTests extends VelocityTestCase
 
         $expected = '#if($map.hasElement($FOO.list(), "BAR", "(?s)\Atrue\Z"))';
 
-        $syntax = $this->translator->createSyntax('ApacheVelocity');
+        $syntax = $this->translator->createApacheVelocity();
+
+        $result = $syntax->translateCommand($cmd);
+
+        $this->assertEquals($expected, $result);
+    }
+
+    public function test_show_decryption()
+    {
+        $cmd = Mailcode_Factory::show()
+            ->var("FOO.BAR")
+            ->enableDecryption();
+
+        $expected = '${text.decrypt($FOO.BAR, "default")}';
+
+        $syntax = $this->translator->createApacheVelocity();
+
+        $result = $syntax->translateCommand($cmd);
+
+        $this->assertEquals($expected, $result);
+    }
+
+    public function test_show_decryption_custom()
+    {
+        $cmd = Mailcode_Factory::show()->var("FOO.BAR")
+            ->enableDecryption('barfoo');
+
+        $expected = '${text.decrypt($FOO.BAR, "barfoo")}';
+
+        $syntax = $this->translator->createApacheVelocity();
 
         $result = $syntax->translateCommand($cmd);
 
@@ -34,7 +63,7 @@ final class Translator_ApacheVelocityTests extends VelocityTestCase
 
     public function test_translateSafeguard()
     {
-        $syntax = $this->translator->createSyntax('ApacheVelocity');
+        $syntax = $this->translator->createApacheVelocity();
 
         $subject = '
 {setvar: $CUSTOMER.CUSTOMER_ID = "42"}
@@ -71,6 +100,32 @@ ${CUSTOMER.CUSTOMER_ID}
         $this->assertEquals($expected, $result);
     }
 
+    public function test_showvar_decrypt() : void
+    {
+        $syntax = $this->translator->createApacheVelocity();
+
+        $subject = '{showvar: $CUSTOMER.CUSTOMER_ID decrypt="my-key" idnencode:}';
+        $expected = '${text.idn(${text.decrypt($CUSTOMER.CUSTOMER_ID, "my-key")})}';
+
+        $safeguard = Mailcode::create()->createSafeguard($subject);
+        $result = $syntax->translateSafeguard($safeguard);
+
+        $this->assertEquals($expected, $result);
+    }
+
+    public function test_showvar_decrypt_custom() : void
+    {
+        $syntax = $this->translator->createApacheVelocity();
+
+        $subject = '{showvar: $CUSTOMER.CUSTOMER_ID decrypt="barfoo" idnencode:}';
+        $expected = '${text.idn(${text.decrypt($CUSTOMER.CUSTOMER_ID, "barfoo")})}';
+
+        $safeguard = Mailcode::create()->createSafeguard($subject);
+        $result = $syntax->translateSafeguard($safeguard);
+
+        $this->assertEquals($expected, $result);
+    }
+
     /**
      * Configuring the translation of showdate commands to use a
      * specific internal date format when they are translated,
@@ -83,7 +138,7 @@ ${CUSTOMER.CUSTOMER_ID}
         $internalFormat = 'yyyy-MM-dd';
         $expected = '${time.input("'.$internalFormat.'", $FOO.BAR).output("dd.MM.yyyy").zone("UTC")}';
 
-        $syntax = $this->translator->createSyntax('ApacheVelocity');
+        $syntax = $this->translator->createApacheVelocity();
         $safeguard = Mailcode::create()->createSafeguard($subject);
         $dateCommands = $safeguard->getCollection()->getShowDateCommands();
 
@@ -103,7 +158,7 @@ ${CUSTOMER.CUSTOMER_ID}
         $internalFormat = 'yyyy-MM-dd';
         $expected = '${esc.url($time.input("' . $internalFormat . '", $FOO.BAR).output("yyyy-MM-dd").zone("UTC"))}';
 
-        $syntax = $this->translator->createSyntax('ApacheVelocity');
+        $syntax = $this->translator->createApacheVelocity();
         $safeguard = Mailcode::create()->createSafeguard($subject);
         $dateCommands = $safeguard->getCollection()->getShowDateCommands();
 
@@ -123,7 +178,7 @@ ${CUSTOMER.CUSTOMER_ID}
         $internalFormat = 'yyyy-MM-dd';
         $expected = '${esc.url($time.input("' . $internalFormat . '", $FOO.BAR).output("yyyy-MM-dd").zone("UTC"))}';
 
-        $syntax = $this->translator->createSyntax('ApacheVelocity');
+        $syntax = $this->translator->createApacheVelocity();
         $safeguard = Mailcode::create()->createSafeguard($subject);
         $dateCommands = $safeguard->getCollection()->getShowDateCommands();
 
@@ -143,7 +198,7 @@ ${CUSTOMER.CUSTOMER_ID}
         $internalFormat = 'yyyy-MM-dd';
         $expected = '${esc.url($time.input("' . $internalFormat . '", $FOO.BAR).output("yyyy-MM-dd").zone("Europe/Berlin"))}';
 
-        $syntax = $this->translator->createSyntax('ApacheVelocity');
+        $syntax = $this->translator->createApacheVelocity();
         $safeguard = Mailcode::create()->createSafeguard($subject);
         $dateCommands = $safeguard->getCollection()->getShowDateCommands();
 
@@ -163,7 +218,7 @@ ${CUSTOMER.CUSTOMER_ID}
         $internalFormat = 'yyyy-MM-dd';
         $expected = '${esc.url($time.input("' . $internalFormat . '", $FOO.BAR).output("yyyy-MM-dd").zone("Europe/Berlin"))}';
 
-        $syntax = $this->translator->createSyntax('ApacheVelocity');
+        $syntax = $this->translator->createApacheVelocity();
         $safeguard = Mailcode::create()->createSafeguard($subject);
         $dateCommands = $safeguard->getCollection()->getShowDateCommands();
 
@@ -183,7 +238,7 @@ ${CUSTOMER.CUSTOMER_ID}
         $internalFormat = 'yyyy-MM-dd';
         $expected = '${esc.url($time.input("' . $internalFormat . '", $FOO.BAR).output("yyyy-MM-dd").zone("Europe/Berlin"))}';
 
-        $syntax = $this->translator->createSyntax('ApacheVelocity');
+        $syntax = $this->translator->createApacheVelocity();
         $safeguard = Mailcode::create()->createSafeguard($subject);
         $dateCommands = $safeguard->getCollection()->getShowDateCommands();